Refactor to aid simplification
