+++ tests/configuration.py
@@ -1,0 +1,3 @@
+import os
+
+PROJECT_ROOT = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))
